+++ backends-gazelle/src/test/resources/tpch-data-parquet/part-00000-9e6b0a02-a781-4a24-bf59-3315473c2bd7-c000.snappy.parquet
