+++ backends-velox/src/main/scala/org/apache/spark/sql/util/ArrowUtils.scala
@@ -19,26 +19,15 @@ package org.apache.spark.sql.util
 
 import scala.collection.JavaConverters._
 
-import org.apache.arrow.memory.RootAllocator
 import org.apache.arrow.vector.complex.MapVector
 import org.apache.arrow.vector.types.{DateUnit, FloatingPointPrecision, TimeUnit}
 import org.apache.arrow.vector.types.pojo.{ArrowType, Field, FieldType, Schema}
-import org.apache.arrow.dataset.file.FileSystemDatasetFactory
-import org.apache.arrow.dataset.file.FileFormat
 
 import org.apache.spark.sql.catalyst.expressions.Attribute
-import org.apache.spark.sql.execution.datasources.v2.arrow.{SparkMemoryUtils, SparkSchemaUtils}
 import org.apache.spark.sql.internal.SQLConf
 import org.apache.spark.sql.types._
-import org.apache.hadoop.fs.FileStatus
-
-import java.net.URI
 
 object ArrowUtils {
-
-  @deprecated
-  val rootAllocator = new RootAllocator(Long.MaxValue)
-
   // todo: support more types.
 
   /** Maps data type from Spark to Arrow. NOTE: timeZoneId required for TimestampTypes */
@@ -161,56 +150,4 @@ object ArrowUtils {
 
   def fromAttributes(attributes: Seq[Attribute]): StructType =
     StructType.fromAttributes(attributes)
-
-  private def rewriteUri(encodeUri: String): String = {
-    val decodedUri = encodeUri
-    val uri = URI.create(decodedUri)
-    if (uri.getScheme == "s3" || uri.getScheme == "s3a") {
-      val s3Rewritten = new URI("s3", uri.getAuthority,
-        uri.getPath, uri.getQuery, uri.getFragment).toString
-      return s3Rewritten
-    }
-    val sch = uri.getScheme match {
-      case "hdfs" => "hdfs"
-      case "file" => "file"
-    }
-    val ssp = uri.getScheme match {
-      case "hdfs" => uri.getSchemeSpecificPart
-      case "file" => "//" + uri.getSchemeSpecificPart
-    }
-    val rewritten = new URI(sch, ssp, uri.getFragment)
-    rewritten.toString
-  }
-
-  def makeArrowDiscovery(encodedUri: String,
-                         startOffset: Long, length: Long): FileSystemDatasetFactory = {
-
-    val format = FileFormat.PARQUET
-    val allocator = SparkMemoryUtils.contextArrowAllocator()
-    val factory = new FileSystemDatasetFactory(allocator,
-      null, // SparkMemoryUtils.contextMemoryPool()
-      format,
-      rewriteUri(encodedUri),
-      startOffset,
-      length)
-    factory
-  }
-
-  def readSchema(file: FileStatus): Option[StructType] = {
-    val factory: FileSystemDatasetFactory =
-      makeArrowDiscovery(file.getPath.toString, -1L, -1L)
-    val schema = factory.inspect()
-    try {
-      Option(SparkSchemaUtils.fromArrowSchema(schema))
-    } finally {
-      factory.close()
-    }
-  }
-
-  def readSchema(files: Seq[FileStatus]): Option[StructType] = {
-    if (files.isEmpty) {
-      throw new IllegalArgumentException("No input file specified")
-    }
-    readSchema(files.toList.head) // todo merge schema
-  }
 }
